# cmd/qris-server/openapi.yaml
openapi: 3.1.0
info:
  title: qris-server
  version: 0.1.0
  description: |
    HTTP API stateless untuk konversi & validasi QRIS. Tanpa database, tanpa
    autentikasi — jalankan di jaringan internal atau di belakang gateway sendiri.

    Sumber: https://github.com/saquone/qris
  license:
    name: MIT
    identifier: MIT

servers:
  - url: http://localhost:8080

paths:
  /health:
    get:
      summary: Cek server hidup
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, const: ok }

  /to-dynamic:
    post:
      summary: QRIS statis → dinamis bernominal
      description: |
        Tag 01 diubah 11 → 12, nominal disisipkan di tag 54, CRC (tag 63) dihitung
        ulang. Identitas merchant tidak disentuh.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payload, amount]
              properties:
                payload:
                  type: string
                  description: Payload QRIS statis
                  example: "00020101021126650014COM.GO-JEK.WWW…63048748"
                amount:
                  type: integer
                  format: int64
                  minimum: 1
                  description: Nominal Rupiah bulat
                  example: 50137
      responses:
        "200":
          description: Payload dinamis
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"

  /validate:
    post:
      summary: Cek payload QRIS statis IDR
      description: |
        Selalu balik 200 — hasilnya ada di field `valid`. Status 400 hanya untuk
        request yang tidak bisa dibaca.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payload]
              properties:
                payload: { type: string }
      responses:
        "200":
          description: Hasil pemeriksaan
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid: { type: boolean }
                  error:
                    type: [string, "null"]
                    description: Alasan bila tidak valid
                    examples: ["QRIS tidak valid (checksum)", "Ini QRIS dinamis, butuh QRIS statis"]
        "400":
          $ref: "#/components/responses/BadRequest"

  /merchant:
    post:
      summary: Nama merchant (tag 59)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payload]
              properties:
                payload: { type: string }
      responses:
        "200":
          description: Nama merchant, string kosong bila payload tak terbaca
          content:
            application/json:
              schema:
                type: object
                properties:
                  name: { type: string, example: "TOKO DEMO" }
        "400":
          $ref: "#/components/responses/BadRequest"

  /decode:
    post:
      summary: Baca payload QR dari gambar
      description: Body = bytes gambar mentah, bukan JSON. Maksimal 5 MB.
      requestBody:
        required: true
        content:
          image/png:
            schema: { type: string, format: binary }
          image/jpeg:
            schema: { type: string, format: binary }
      responses:
        "200":
          description: Payload QR
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"

  /qris:
    post:
      summary: Unggah QRIS statis
      description: |
        Body = bytes gambar mentah. Payload di-decode, divalidasi sebagai QRIS statis IDR,
        gambarnya disimpan di folder `-qris-dir` dan payload teksnya di SQLite.
      requestBody:
        required: true
        content:
          image/png: { schema: { type: string, format: binary } }
          image/jpeg: { schema: { type: string, format: binary } }
      responses:
        "200":
          description: Merchant tersimpan
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  name: { type: string, example: "TOKO DEMO" }
                  image_file: { type: string }
                  payload: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
    get:
      summary: Daftar QRIS statis tersimpan
      responses:
        "200": { description: OK }

  /charges:
    post:
      summary: Buat tagihan (QRIS statis → dinamis bernominal)
      description: |
        Kode unik 1..999 ditambahkan ke nominal supaya dua tagihan tidak pernah bernominal
        sama — tanpa itu notifikasi masuk tidak bisa dipetakan ke satu tagihan tertentu.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                merchant_id: { type: integer, description: "kosong = QRIS pertama" }
                amount: { type: integer, format: int64, example: 50000 }
                unique_min: { type: integer, default: 1 }
                unique_max: { type: integer, default: 999 }
                expires_in_seconds: { type: integer, default: 900 }
      responses:
        "200":
          description: Tagihan dibuat
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  base_amount: { type: integer, format: int64 }
                  unique_code: { type: integer }
                  amount: { type: integer, format: int64, description: "nominal final yang harus dibayar" }
                  payload: { type: string, description: "QRIS dinamis" }
                  status: { type: string, enum: [pending, paid, expired] }
        "400": { $ref: "#/components/responses/BadRequest" }
    get:
      summary: Daftar tagihan
      responses:
        "200": { description: Terbaru lebih dulu }

  /charges/{id}:
    get:
      summary: Status satu tagihan
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: OK }
        "400": { $ref: "#/components/responses/BadRequest" }

  /gateways:
    get:
      summary: Katalog gateway yang didukung
      description: |
        Satu-satunya sumber kebenaran tentang aplikasi mana yang didukung dan pola apa yang
        dipakai membaca nominalnya. Aplikasi Android mengambil daftar ini lalu menyimpannya
        untuk dipakai offline.
      responses:
        "200":
          description: Daftar gateway
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    key: { type: string, example: "gopay" }
                    label: { type: string, example: "GoPay Merchant" }
                    packages:
                      type: array
                      items: { type: string }
                      description: Nama paket Android
                    patterns:
                      type: array
                      items: { type: string }
                      description: Pola regex, dicoba berurutan; tiap pola satu capture group

  /notification:
    post:
      summary: Terima notifikasi dari aplikasi listener Android
      description: |
        Menerima langsung payload dari
        [android-notification-listener](https://github.com/Saquone/android-notification-listener):
        verifikasi tanda tangan, ekstrak nominal, kembalikan hasilnya.

        Pola nominal diambil dari katalog bawaan, atau dari berkas `-patterns` bila diberikan.
        Bila `-secret` diberikan, header `X-Signature` (HMAC-SHA256 hex atas body) wajib cocok.

        **Selalu 200 saat tanda tangan sah**, termasuk ketika nominal tidak terbaca
        (`matched: false`) — notifikasi promo/cashback ikut terkirim, dan menjawabnya dengan
        error membuat aplikasi mengirim ulang selamanya.
      parameters:
        - in: header
          name: X-Signature
          schema: { type: string }
          description: HMAC-SHA256 hex atas body. Wajib bila server dijalankan dengan `-secret`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [package_name, title, text, posted_at]
              properties:
                package_name: { type: string, example: "id.dana" }
                title: { type: string, example: "Pembayaran Masuk" }
                text: { type: string, example: "Rp50.137 diterima DANA Bisnis." }
                posted_at:
                  type: integer
                  format: int64
                  description: Epoch milidetik
                  example: 1765432100000
      responses:
        "200":
          description: Hasil ekstraksi
          content:
            application/json:
              schema:
                type: object
                properties:
                  package_name: { type: string }
                  posted_at: { type: integer, format: int64 }
                  matched: { type: boolean, description: "nominal berhasil dibaca" }
                  amount:
                    type: [integer, "null"]
                    description: Nominal Rupiah bulat, `null` bila tidak terbaca
                  verified:
                    type: boolean
                    description: Nominal cocok persis dengan tagihan pending, dan tagihan itu ditandai lunas
                  charge_id:
                    type: [integer, "null"]
                    description: Tagihan yang dilunasi notifikasi ini
        "401":
          description: Tanda tangan tidak cocok
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"

  /notifications:
    get:
      summary: Daftar notifikasi tersimpan
      description: Aktif bila server dijalankan dengan `-db` (default `qris.db`).
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 50, minimum: 1, maximum: 500 }
      responses:
        "200":
          description: Terbaru lebih dulu
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: integer }
                        package_name: { type: string }
                        title: { type: string }
                        text: { type: string }
                        posted_at: { type: integer, format: int64 }
                        amount: { type: [integer, "null"] }
                        received_at: { type: integer, format: int64 }

  /parse:
    post:
      summary: Ekstrak nominal dari teks notifikasi
      description: |
        Pola dicoba berurutan; yang pertama menghasilkan nominal valid menang.
        Tiap pola wajib punya satu capture group berisi angka.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patterns, text]
              properties:
                patterns:
                  type: array
                  items: { type: string }
                  example: ["menerima Rp ?([0-9.,]+)"]
                text:
                  type: string
                  example: "DANA: Anda menerima Rp 50.137 dari BUDI"
      responses:
        "200":
          description: Nominal Rupiah bulat
          content:
            application/json:
              schema:
                type: object
                properties:
                  amount: { type: integer, format: int64, example: 50137 }
        "400":
          $ref: "#/components/responses/BadRequest"

components:
  responses:
    BadRequest:
      description: Request tidak valid
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: "nominal tidak valid" }